add tests for group data_array and tag access
